Fix notifications
expo-notifications needed a native release. Releasing react native bundle
for devices without latest version of native app caused app breakage. To fix
this adding a condition to use expo-notifications only on the latest version
of the native app (1.0.5 iOS and 12 android version code)

TODO - this condition should be removed once all users are on the above
version of the app